+++ engine/src/libCoreServiceKit/core_connection_to_service.cpp
@@ -69,7 +69,7 @@ namespace core
 			CCoreServiceKitImpl::Inst()->getCoreServiceProxy()->delConnectionToService(this->m_szServiceName);
 	}
 
-	void CCoreConnectionToService::onDispatch(uint32_t nMessageType, const void* pData, uint16_t nSize)
+	void CCoreConnectionToService::onDispatch(uint8_t nMessageType, const void* pData, uint16_t nSize)
 	{
 		if (nMessageType == eMT_SYSTEM)
 		{
